move messages to new file + stuff
